Split server.py — extract projects route domain (#590)
Move the 7 projects-domain handlers (api_projects, api_projects_create,
api_projects_delete, api_roles, api_session_defaults, api_check_path,
api_check_branches) plus the projects-private _scan_machine_projects helper
verbatim into agentwire/routes/projects.py as ProjectsRoutesMixin +
register_projects_routes, following the #584 mixin/registrar seam.

server.py: add the mixin to AgentWireServer bases, swap the inline projects
add_* calls for register_projects_routes(), delete the moved handlers.
api_machine_status stays on base (machines domain, #589).

Pure relocation — route set invariant. Route-snapshot guard + full
unit+integration suite green (2134 passed); portal boots, all 7 routes 200.

Closes #590

Built by dotdev.dev
